Hero will now regenerate health over time. Six hours after his last death, he is all filled up again.
//TODO: Actually test that. But 6 hours is looong.

Experience to next level is now active.

Prices will adjust according to the hero's charisma.
Rebatefunction in statcalculator

TODO: Change name of statCalculator

There is now a field outside the town. It will generate 5 enemies, and send the hero against a random amount of them.
Usually ten times before he can come back, but it is adjustable.

TODO: Test this as well

Cap of numbergeneration was increased. We always use modulo anyways.